requirements.txt
Flask==3.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==42.0.5
python-dotenv==1.0.1
flake8==7.1.1
gunicorn==23.0.0
requests==2.31.0